refactor(protocol): drop hand-rolled RecipeDisplay/SlotDisplay/RecipeBookEntry
Now that the codegen handles these natively, delete the hand-rolled
versions from `crates/basalt-protocol/src/types/recipe_display.rs`
(775 lines). Move the surviving exception — `IDSet` — into a focused
`id_set.rs` module with a top-level doc comment explaining why this
single type stays manual: `registryEntryHolderSet` encodes the
variant count in the discriminator tag itself (`tag = N + 1` for `N`
inline varints), a "tag-is-a-count" pattern no other Mojang type
uses. Generalising it would be more code than it saves.

Add `crates/basalt-protocol/tests/recipe_display_codegen.rs` with
13 round-trip tests covering every `RecipeDisplay` and `SlotDisplay`
variant — including recursive cases (`SmithingTrim`, `WithRemainder`,
`Composite`) — plus the `crafting_requirements: Option<Vec<u8>>`
bool-prefix encoding for both `None` and `Some(payload)`. Byte
fixtures are byte-equivalent to the deleted hand-rolled tests so
the wire format is provably unchanged after the migration.
